organize auth plugins into per-plugin folders
