+++ app/allowlist_test.go
@@ -5,8 +5,9 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	_ "github.com/winhowes/AuthTransformer/app/authplugins/incoming"
-	_ "github.com/winhowes/AuthTransformer/app/authplugins/outgoing"
+	_ "github.com/winhowes/AuthTransformer/app/authplugins/basic"
+	_ "github.com/winhowes/AuthTransformer/app/authplugins/google_oidc"
+	_ "github.com/winhowes/AuthTransformer/app/authplugins/token"
 	_ "github.com/winhowes/AuthTransformer/app/secrets/plugins"
 )
