+++ app/integration_test.go
@@ -3,8 +3,9 @@ package main
 import (
 	"testing"
 
-	_ "github.com/winhowes/AuthTransformer/app/authplugins/incoming"
-	_ "github.com/winhowes/AuthTransformer/app/authplugins/outgoing"
+	_ "github.com/winhowes/AuthTransformer/app/authplugins/basic"
+	_ "github.com/winhowes/AuthTransformer/app/authplugins/google_oidc"
+	_ "github.com/winhowes/AuthTransformer/app/authplugins/token"
 	_ "github.com/winhowes/AuthTransformer/app/secrets/plugins"
 )
