+++ app/main.go
@@ -16,8 +16,9 @@ import (
 	"time"
 
 	"github.com/winhowes/AuthTransformer/app/authplugins"
-	_ "github.com/winhowes/AuthTransformer/app/authplugins/incoming"
-	_ "github.com/winhowes/AuthTransformer/app/authplugins/outgoing"
+	_ "github.com/winhowes/AuthTransformer/app/authplugins/basic"
+	_ "github.com/winhowes/AuthTransformer/app/authplugins/google_oidc"
+	_ "github.com/winhowes/AuthTransformer/app/authplugins/token"
 	_ "github.com/winhowes/AuthTransformer/app/secrets/plugins"
 )
